refactor discovery logic into separate files from network.c/h
